refactor: restructure the repository
